Module Alba_core

module Sort : sig ... end

module Term : sig ... end

module Inductive : sig ... end

module Operator : sig ... end

module Gamma : sig ... end

module Typecheck : sig ... end

module Name_map : sig ... end

module Context : sig ... end

module Unifier : sig ... end

module Gamma_holes : sig ... end

Gamma_holes is a context with holes which can be filled later. A hole is a local unnamed variable with a type (i.e. an assumption that an element with this type exists), initially without value. Later on the value can be provided.

module Gamma_algo : sig ... end

module Term_printer : sig ... end

module Print_inductive : sig ... end