Module Term_printer.Pretty

Pretty printer which can print terms in a context with the use of apretty printer.

Parameters

G : GAMMA

P : Fmlib.Pretty_printer.SIG

Signature

val print : Term.t -> G.t -> P.t

print term gamma Print term in the context gamma using the printer P.