Module Gamma_algo.Make

Parameters

Gamma : GAMMA

Signature

val type_of_term : Alba_core.Term.t -> Gamma.t -> Alba_core.Term.typ

type_of_term term gamma

Compute the type of term

Precondition: term must be welltyped and the context valid.

val split_type : Alba_core.Term.typ -> Gamma.t -> (Alba_core.Term.Pi_info.t * Alba_core.Term.typ) list * Alba_core.Term.typ

val split_kind : Alba_core.Term.typ -> Gamma.t -> ((Alba_core.Term.Pi_info.t * Alba_core.Term.typ) list * Alba_core.Sort.t) option

split_kind k gamma:

Compute the arguments and the sort of the kind k. If typ does not reduce to a kind, then return None.

Precondition: k must be welltyped and the context gamma must be valid.

A kind has the form

all (x: A) (y: B) .... : s

where s is a sort.

val sort_of_kind : Alba_core.Term.typ -> Gamma.t -> Alba_core.Sort.t option

sort_of_kind typ gamma

Compute the sort of the kind typ. If typ does not reduce to a kind, then return None.

Precondition: typ must be welltyped and the context gamma must be valid.

A kind has the form

all (x: A) (y: B) .... : s

where s is a sort.

val is_kind : Alba_core.Term.typ -> Gamma.t -> bool

is_kind typ gamma

Is the welltyped term typ in the valid context gamma a kind?

val key_split : Alba_core.Term.t -> Gamma.t -> Alba_core.Term.t * (Alba_core.Term.t * Alba_core.Term.Application_info.t) list

val key_normal : Alba_core.Term.t -> Gamma.t -> Alba_core.Term.t

keynormal term gamma

Compute the key normal form of the welltyped term term in the valid context gamma.

val normalize_pi : Alba_core.Term.typ -> Gamma.t -> Alba_core.Term.typ

normalize_pi typ gamma

Precondition:

typ must be a valid type

Result: typ is expanded until the form

all (a: A) (b: B) ... : R

is reached where R cannot be expanded further into a product type.

val normalize : Alba_core.Term.t -> Gamma.t -> Alba_core.Term.t

normalize term gamma

Compute the normal form of the welltyped term term in the valid context gamma.

val check_naming_convention : string -> Alba_core.Term.typ -> Gamma.t -> (unit, name_violation) Stdlib.result