Parameter Make.1-Gamma

type t

val count : t -> int

Number of variables in the context.

val is_valid_index : int -> t -> bool

val name_of_index : int -> t -> string

val push_local : string -> Alba_core.Term.typ -> t -> t

push_local name typ gamma

Add the local variable with name and typ to the context.

Precondition

name <> ""

val type_of_literal : Alba_core__.Value.t -> t -> Alba_core.Term.typ

val type_of_variable : int -> t -> Alba_core.Term.typ

type_of_variable idx gamma

Return the type of the variable idx. idx is the De Bruijn index i.e. and index of 0 points to the last variable and an index of count gamma - 1 points to the first variable.

val definition_term : int -> t -> Alba_core.Term.t option

definition_term idx gamma

If the variable idx has a definition, then return it. Otherwise return None.