Module Welltyped.Builder

A builder for welltyped terms in wellformed contexts.

Parameters

Info : Fmlib.Module_types.ANY

Signature

type problem = Info.t * Alba_core__.Type_error.t

type 'a res = ('a, problem) Stdlib.result

'a res The result of a building process.

type t

type name = Info.t * string

type formal_argument = name * t

type signature = formal_argument list * t

val sort : Info.t -> Alba_core.Sort.t -> t

val variable : Info.t -> int -> t

val identifier : Info.t -> string -> t

identifier info name Build the term represented by name.

val unknown : Info.t -> t

Unknown term. The compiler is asked to derive.

val application : Info.t -> t -> t -> t

val lambda : Info.t -> name -> t -> t -> t

lambda name typ exp Build the lambda term \ (name: typ) := exp.

val pi : Info.t -> name -> t -> t -> t

pi name typ res Build the product all (name: typ): res.

val make_term : context -> t -> judgement res

make_term c term Build the term term in context c.

val make_builtin : context -> name -> signature -> context res

val make_definition : Info.t -> name -> signature -> t -> context -> context res