Module Welltyped.Check

A type checker.

type 'a res = ('a, Alba_core__.Type_error.t) Stdlib.result

val check_term : Alba_core.Term.t -> context -> judgement res

check term context Check term in the wellformed context and return a judgement, i.e. a welltyped term with its type in the same context or return a type error.