Module Alba_core__Term_printer

module type GAMMA = sig ... end

Signature which must be satisfied by a valid context.

module Pretty : functor (G : GAMMA) -> functor (P : Fmlib.Pretty_printer.SIG) -> sig ... end

Pretty printer which can print terms in a context with the use of apretty printer.

module String_print : functor (G : GAMMA) -> sig ... end

Pretty printer which prints terms in a context into a string.

val string_of_term : Alba_core.Term.t -> Alba_core.Gamma.t -> string