Module Alba_core__Term_printer.String_print

Pretty printer which prints terms in a context into a string.

Parameters

G : GAMMA

Signature

val string_of_term : Alba_core.Term.t -> G.t -> string

string_of_term term gamma. Pretty print term in the context gamma into a string.