Module Alba_core__Unifier.Make

Parameters

H : HOLES

Signature

val unify : Alba_core.Term.typ -> Alba_core.Term.typ -> bool -> H.t -> H.t option

unify act req is_super gh

Unify the type act with the type req using gh as the context with holes.

is_super indicates if the typ req can be regarded as a supertype of act.