Module 1-IO.Path

val absolute : string -> string t

absolute path converts path into an absolute path.

val split : string -> (string * string) option

split path splits path into a dirname and a basename if possible.

Examples:

split ""                = None
split "/"               = None
split "/hello"          = Some ("/", "hello")
split "/User/name/xxx/" = Some("/User/name", "xxx")
split "/User/name/xxx"  = Some("/User/name", "xxx")

val normalize : string -> string

normalize path removes duplicate path separators and normalizes "." and ".." segments.

Examples:

normalize ""            = "."
normalize "/"           = "/"
normalize "////"        = "/"
normalize "a//b"        = "a/b"
normalize "a/./b"       = "a/b"
normalize "a/../b"      = "b"
normalize "a/b/../../c" = "c"
normalize "../a"        = "../a"

val join : string -> string -> string

join dir file joins the directory name dir with the file name file.