Module Albalib__Build_context.Application

val start : int -> t -> t

start nargs bc Start a function application with nargs arguments.

val apply : int -> Alba_core.Term.Application_info.t -> t -> (t, type_in_context * type_in_context) Stdlib.result

apply n_remaining mode bc Apply the function to the argument. There are n_remaining remaining arguments.