Module Albalib__Build_context.Lambda

val start : t -> t

Start a function abstraction and expect the first argument type.

val inner : t -> t

Expect the inner expression of the function abstraction whose type is the last analyzed expression.

val end_ : int -> int -> bool -> t -> (t, type_in_context * type_in_context) Stdlib.result

end_ nargs nbounds typed cb

End the function abstraction with nbounds bound variables. The function abstraction is applied to nargs arguments.