Module Fmlib

Fmlib: Functional Monadic Library

Parsing

module Position : sig ... end

Represents a position in a text file.

module Character_parser : sig ... end

module Generic_parser : sig ... end

Pretty Printing

module Pretty_print : sig ... end

Pretty Printer: Generate nicely formatted ascii text.

module Pretty_printer : sig ... end

Old pretty printer, deprecated

Applications

Console Applications

module Io : sig ... end

Definition of an enviroment for console applications which can run natively or under nodejs or any other module which satisfies the signature SIG.

module Make_io : sig ... end

This module contains a minimal signature for an io module and a functor which converts a minimal implementation into a full implementation of an io module.

module String_printer : sig ... end

Web Applications

module Web_application : sig ... end

Enviroment for Web Applications.

Basic Functions and Data Types

val identity : 'a -> 'a

Identity function

module Module_types : sig ... end

Common module types like ANY, SORTABLE, ...

module Array : sig ... end

module List : sig ... end

A thin wrapper around Stdlib.List which avoids throwing exceptions and with some additional monadic functions.

module Deque : sig ... end

module Finite_map : sig ... end

Thin wrapper around Stdlib.Map which supports monadic operations instead of throwing exceptions.

module Monad : sig ... end

module Option : sig ... end

Option Monad

module Pool : sig ... end

An arrayed pool of elements of a certain type.

module Red_black : sig ... end

Finite maps and sets based on red black trees.

module Readable_printer : sig ... end

module Result : sig ... end

Result Monad

module Sequence : sig ... end

module Vector : sig ... end

A mutable vector (like a C++ vector)

module Common : sig ... end

Old modules (deprecated)

module Argument_parser : sig ... end