Module Common.String

type t = string

val is_prefix : t -> t -> bool

is_prefix a b Is a a prefix of b?

val is_suffix : t -> t -> bool

is_suffix a b Is a a suffix of b?

val reverse : t -> t

Reverse the string.

val compare : t -> t -> int

val one : char -> t

val find : (char -> bool) -> int -> t -> int

val has : (char -> bool) -> int -> t -> bool

val find_bwd : (char -> bool) -> int -> t -> int

val list : t -> char list

val of_list : char list -> t

val length : t -> int

val get : t -> int -> char

val escaped : t -> t

val sub : t -> int -> int -> t

val concat : string -> string list -> string

val split_on_char : char -> string -> string list

val make : int -> char -> t

val init : int -> (int -> char) -> t

module To_readable : sig ... end

module From_readable : functor (R : Module_types.READABLE) -> sig ... end