Module Common.String_reader

include Fmlib.Module_types.READABLE

type t

Type of the structure.

val has_more : t -> bool

Does the structure have more characters to read?

val peek : t -> char

peek r returns the next character.

val advance : t -> t

advance r advances the structure by one character.

val of_string : string -> t

of_string s creates a readable structure of the string s.

val of_substring : string -> int -> int -> t

of_substring s start len creates a readable structure of the substring of s starting at position start and having length len.