Module Fmlib__.Make_io

module type SIG = sig ... end

Minimal signature, which an implementation of an Io module must satisfy.

module Make : functor (Base : SIG) -> Fmlib.Io.SIG

Make (Base) transforms a basic implementation of an io module which conforms to the signature SIG into a full implementation of an io module.