Module Fmlib__.Readable_printer

Principles

Readable Structure

module R : Fmlib.Module_types.READABLE

The readable structure returned by the printer.

Basics

type t

The type of a readable printer.

val empty : t

The readable printer printing nothing.

val (<+>) : t -> t -> t

p1 <+> p2 prints first p1 and then p2.

Print Commands

val string : string -> t

string s prints the string s.

val substring : string -> int -> int -> t

substring s start len prints the substring of s starting at position start with length len.

val fill : int -> char -> t

fill n c prints the character c n times.

val char : char -> t

char c prints the character c.

Generate the Readable Structure

val readable : t -> R.t

readable p generates a readable structure from the printer p.