Parameter Read.1-W

type t

Type of the structure.

val needs_more : t -> bool

Is it possible to write more characters to the structure?

val put_character : t -> char -> t

put_character w c writes the character c to the structure and returns a structure which might accept more characters.

val put_end : t -> t

put_end w signals to the structure w that there are no more characters available to write (e.g. eof reached).