Module Fmlib__Position.Print

Print in memory source files with error markers.

Parameters

PP : Fmlib.Pretty_printer.SIG

Signature

val print_source_lines : string Fmlib.Sequence.t -> range -> PP.t

print_source_lines lines range Print the source file given as a sequence of lines with line numbers and highlight the region range.

val print_source : string -> range -> PP.t

print_source src range Print the source file given as a string with line numbers and highlight the region range.