Module 1-App.Command

type 'msg t =

| None

| Batch of 'msg t list

| Http of string * string * string * (string, int) Stdlib.result -> 'msg
Http (type, url, data, handler)

The type is GET, POST, HEAD, etc.

data will be sent as part of the request (makes sense only for POST and related request types).

handler must be able to react to the response text or the status in case an error occurs (the famous 404 for "page not found").